feat(api): ajout des tables de la BDD
* build(api): add lombok dependency

* feat(api): ajout de toutes les tables à shema.sql

* feat(api-controller,dao): ajout de la base de la BDD fonctionnel

* fix(api-model): change les table aux noms reservés USER et ORDER

+++ api/src/main/java/fr/epf/speedycart/api/dao/ClientDao.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface ClientDao extends JpaRepository<Client, Integer> {
-    Client findById(int id);
+    Client findById(long id);
     List<Client> findAll();
 }
