+++ api/src/main/java/fr/epf/speedycart/api/dao/ProductDao.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface ProductDao extends JpaRepository<Product, Integer> {
-    Product findById(int id);
+    Product findById(long id);
     List<Product> findAll();
 }
